refactor: bottomSheet 해당하는 파일 패키지 재수정
- 스크롤 가능한 아이템 영역들이 bottomSheet 스크롤과 어색한 흐름으로 작동해서 오버레이 방식으로 변경하며 패키지도 이에 맞게 재수정
