+++ android/app/src/main/java/com/on/turip/ui/compose/trip/turipselection/component/TuripDetail.kt
@@ -1,4 +1,4 @@
-package com.on.turip.ui.compose.trip.bottomsheet.component
+package com.on.turip.ui.compose.trip.turipselection.component
 
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.Image
@@ -38,10 +38,10 @@ import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
 import com.on.turip.R
 import com.on.turip.ui.compose.designsystem.theme.TuripTheme
-import com.on.turip.ui.compose.trip.bottomsheet.model.TuripPlaceModel
-import com.on.turip.ui.compose.trip.bottomsheet.util.reorderable.ReorderableItem
-import com.on.turip.ui.compose.trip.bottomsheet.util.reorderable.rememberReorderableLazyColumnState
 import com.on.turip.ui.compose.trip.model.MapModel
+import com.on.turip.ui.compose.trip.turipselection.model.TuripPlaceModel
+import com.on.turip.ui.compose.trip.turipselection.util.reorderable.ReorderableItem
+import com.on.turip.ui.compose.trip.turipselection.util.reorderable.rememberReorderableLazyColumnState
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toPersistentList
