+++ android/app/src/main/java/com/on/turip/ui/compose/trip/turipselection/component/TuripItem.kt
@@ -1,4 +1,4 @@
-package com.on.turip.ui.compose.trip.bottomsheet.component
+package com.on.turip.ui.compose.trip.turipselection.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
