+++ android/app/src/main/java/com/on/turip/ui/compose/trip/turipselection/model/DeletePlaceSnapshot.kt
@@ -1,4 +1,4 @@
-package com.on.turip.ui.compose.trip.bottomsheet.model
+package com.on.turip.ui.compose.trip.turipselection.model
 
 import com.google.android.gms.maps.model.LatLng
 import kotlinx.collections.immutable.ImmutableList
