+++ android/app/src/main/java/com/on/turip/ui/main/favorite/TuripMapper.kt
@@ -4,7 +4,7 @@ import androidx.core.net.toUri
 import com.google.android.gms.maps.model.LatLng
 import com.on.turip.domain.bookmark.TuripPlace
 import com.on.turip.domain.turip.Turip
-import com.on.turip.ui.compose.trip.bottomsheet.model.TuripPlaceModel
+import com.on.turip.ui.compose.trip.turipselection.model.TuripPlaceModel
 import com.on.turip.ui.main.favorite.model.PlaceLatLngUiModel
 import com.on.turip.ui.main.favorite.model.TuripModel
 import com.on.turip.ui.main.favorite.model.TuripPlaceShareModel
